% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.simlr.R
\name{adjust_evals}
\alias{adjust_evals}
\title{Adjust eigenvalues (down-weight small values).}
\usage{
adjust_evals(d, alpha = 0.8, beta = 2)
}
\description{
Adjust eigenvalues (down-weight small values).
}
\keyword{internal}
