% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.simlr.R
\name{approx_spaced_integers}
\alias{approx_spaced_integers}
\title{Return a vector of integers from from to to that are approximately evenly spaced}
\usage{
approx_spaced_integers(from, to, max_length)
}
\description{
Return a vector of integers from from to to that are approximately evenly spaced
}
\keyword{internal}
