% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BuettnerFlorian}
\alias{BuettnerFlorian}
\title{test dataset}
\format{gene expression measurements of individual cells}
\source{
Buettner, Florian, et al. "Computational analysis of cell-to-cell heterogeneity in single-cell RNA-sequencing data reveals hidden subpopulations of cells." Nature biotechnology 33.2 (2015): 155-160.
}
\usage{
data(BuettnerFlorian)
}
\value{
list of 6
	in_X = input dataset 
 	n_clust = number of clusters, 
 	true_labs = ground true of cluster assignments, 
 	seed = seed,
 	results = last result of SIMLR,
 	nmi = normalized mutual information
}
\description{
example dataset to test SIMRL from the work by Buettner, Florian, et al.
}

