% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.DD.R
\name{calc.DD}
\alias{calc.DD}
\title{Calculate DD in parallel if cl is not NULL, in serial if not}
\usage{
calc.DD(cl, D_Kernels, S)
}
\description{
Calculate DD in parallel if cl is not NULL, in serial if not
}
\keyword{internal}
