% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.simlr.R
\name{calc_eigs}
\alias{calc_eigs}
\title{Calculate eigenvalues of a symmetric matrix P.}
\usage{
calc_eigs(P, method = "eigen")
}
\description{
Calculate eigenvalues of a symmetric matrix P.
}
\keyword{internal}
