% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMLR.R
\name{CIMLR}
\alias{CIMLR}
\title{CIMLR, refer to https://github.com/danro9685/CIMLR}
\usage{
CIMLR(X, c, no.dim = NA, k = 10, cores.ratio = 1)
}
\arguments{
\item{X}{a list of multi-omic data each of which is an (m x n) data matrix of measurements of cancer patients}

\item{c}{number of clusters to be estimated over X}

\item{no.dim}{number of dimensions}

\item{k}{tuning parameter}

\item{cores.ratio}{ratio of the number of cores to be used when computing the multi-kernel}
}
\value{
clusters the patients based on CIMLR and their similarities

list of 8 elements describing the clusters obtained by CIMLR, of which y are the resulting clusters: 
     y = results of k-means clusterings,
     S = similarities computed by CIMLR, 
     F = results from network diffiusion, 
     ydata = data referring the the results by k-means,
     alphaK = clustering coefficients,
     execution.time = execution time of the present run,
     converge = iterative convergence values by T-SNE,
     LF = parameters of the clustering
}
\description{
perform the CIMLR clustering algorithm
}
