% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.CIMLR.R
\name{CIMLR_Estimate_Number_of_Clusters}
\alias{CIMLR_Estimate_Number_of_Clusters}
\title{This function has been removed and it is now provided
by the CIMLR package, please refer to https://github.com/danro9685/CIMLR}
\usage{
CIMLR_Estimate_Number_of_Clusters(all_data, NUMC = 2:5, cores.ratio = 1)
}
\arguments{
\item{all_data}{is a list of multi-omic data each of which is an (m x n) data matrix of measurements of cancer patients}

\item{NUMC}{vector of number of clusters to be considered}

\item{cores.ratio}{ratio of the number of cores to be used when computing the multi-kernel}
}
\value{
a list of 2 elements: K1 and K2 with an estimation of the best clusters (the lower   
values the better) as discussed in the original paper of SIMLR
}
\description{
CIMLR_Estimate_Number_of_Clusters
}
\examples{
   
\dontrun{
data(GliomasReduced)    
CIMLR_Estimate_Number_of_Clusters(GliomasReduced$in_X,   
   NUMC = 2:5,   
   cores.ratio = 0)  
}
}
