% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.simlr.R
\name{cores_from_ratio}
\alias{cores_from_ratio}
\title{Calculate how many cores from the ratio.}
\usage{
cores_from_ratio(cores.ratio)
}
\description{
Calculate how many cores from the ratio.
}
\keyword{internal}
