% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.simlr.R
\docType{data}
\name{CumulativeTimer}
\alias{CumulativeTimer}
\title{A class to measure cumulative time across sub-tasks.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CumulativeTimer
}
\description{
A class to measure cumulative time across sub-tasks.
}
\keyword{internal}
