% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.diffusion.R
\name{dn}
\alias{dn}
\title{Normalizes a symmetric kernel W.}
\usage{
dn(W, type)
}
\arguments{
\item{:}{W The symmetric kernel}

\item{:}{type The normalisation type
\enumerate{
  \item 'ave' Scales the rows by the column sums.
  \item 'gph' Scales each element by the square root of its column and row sums.
    This is probably for the case when W is not symmetric but does not seem to be used.
}}
}
\description{
That is if W is a symmetric adjacency matrix and D is the corresponding diagonal degree matrix,
compute inv(D) %*% W, which in spectral graph theory is the diffusion operator.
}
\keyword{internal}
