% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.simlr.R
\name{eig1}
\alias{eig1}
\title{Compute the eigenvalues and eigenvectors of A}
\usage{
eig1(A, c = NA, isMax = NA, isSym = NA)
}
\arguments{
\item{A}{The matrix to compute the eigenvalues and vectors of}

\item{c}{The number of eigenvectors to compute}

\item{isMax}{Either return the largest or the smallest eigenvalues (and associated vectors)}

\item{isSym}{Is A symmetric?}
}
\description{
Compute the eigenvalues and eigenvectors of A
}
\keyword{internal}
