% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GliomasReduced}
\alias{GliomasReduced}
\title{test dataset for CIMLR}
\format{multi-omic data of cancer patients}
\source{
Cancer Genome Atlas Research Network. "Comprehensive, integrative genomic analysis of diffuse lower-grade gliomas." New England Journal of Medicine 372.26 (2015): 2481-2498.
}
\usage{
data(GliomasReduced)
}
\value{
list of 1 element: 
	in_X = input dataset as a list of 4 (reduced) multi-omic data each of which is an (m x n) measurements of cancer patients
}
\description{
example dataset to test CIMLR. This is a reduced version of the dataset from the work by The Cancer Genome Atlas Research Network.
}
