% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.multiple.kernel.R
\name{kernel.distance.2}
\alias{kernel.distance.2}
\title{Calculate the kernel distance}
\usage{
kernel.distance.2(G)
}
\arguments{
\item{G}{The kernel (Gram) matrix}
}
\description{
The kernel distance is the distance in feature space. This function
returns half the squared kernel distance for some reason.
}
\keyword{internal}
