% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.simlr.R
\name{kernel_param_map}
\alias{kernel_param_map}
\title{Return a data frame containing the kernel parameters.}
\usage{
kernel_param_map()
}
\description{
Return a data frame containing the kernel parameters.
}
\keyword{internal}
