% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kolodziejczyk}
\alias{Kolodziejczyk}
\title{Data from Kolodziejczyk et al.}
\format{gene expression measurements of individual cells}
\usage{
data(Kolodziejczyk)
}
\value{
list of 6:
  \enumerate{
    \item in_X = input dataset as an (m x n) gene expression measurements of individual cells,
    \item n_clust = number of clusters (number of distinct true labels),
    \item true_labs = ground true of cluster assignments for each of the n_clust clusters,
    \item seed = seed used to compute the results for the example,
    \item results = result by SIMLR for the inputs defined as described,
    \item nmi = normalized mutual information as a measure of the inferred clusters compared to the true labels
  }
}
\description{
Data from Kolodziejczyk et al.
}
