% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.simlr.R
\name{L2_distance_1}
\alias{L2_distance_1}
\title{Compute the L2 distances between the rows of two matrices.}
\usage{
L2_distance_1(a, b = NULL)
}
\arguments{
\item{a:}{First matrix}

\item{b:}{Second matrix (same shape as a). If NULL will be taken as equivalent to a.}
}
\description{
Might as well use `as.matrix(dist(a))^2`
}
\keyword{internal}
