% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.simlr.R
\name{laplacian}
\alias{laplacian}
\title{Calculate Laplacian matrix.}
\usage{
laplacian(S, method = "orig")
}
\description{
Calculate the Laplacian matrix of the graph represented by the adjacency matrix S
https://en.wikipedia.org/wiki/Laplacian_matrix
}
\keyword{internal}
