% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.multiple.kernel.R
\name{multiple.unnorm.kernels}
\alias{multiple.unnorm.kernels}
\title{Compute and returns multiple unnormalised kernels}
\usage{
multiple.unnorm.kernels(x, cores.ratio = 1)
}
\arguments{
\item{x}{The data (samples x features)}

\item{cores.ratio}{Proportional of all possible cores - 1 to use.}
}
\description{
Compute the kernels for the hard-wired range sigma (bandwidth scaling) and k (nearest-neighbours)
}
\keyword{internal}
