% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.diffusion.R
\name{network.diffusion}
\alias{network.diffusion}
\title{Perform network diffusion of K steps over the network A}
\usage{
network.diffusion(A, K)
}
\arguments{
\item{A:}{The distance matrix}

\item{K:}{The number of nearest neighbours to consider}
}
\description{
This is called similarity enhancement by diffusion in the paper
}
