% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.multiple.kernel.R
\name{norm.and.calc.dists}
\alias{norm.and.calc.dists}
\title{Normalise the kernels and convert to kernel distances}
\usage{
norm.and.calc.dists(D_Kernels, offset = 1)
}
\arguments{
\item{D_Kernels}{The unnormalised kernel (Gram) matrices}
}
\description{
Normalise the kernels and convert to kernel distances
}
\keyword{internal}
