% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{similarity.heatmap}
\alias{similarity.heatmap}
\title{Plot a similarity matrix heatmap}
\usage{
similarity.heatmap(S, samples = NULL, label = NULL, cluster = NULL,
  color = colorRampPalette(rev(RColorBrewer::brewer.pal(n = 7, name =
  "RdYlBu")))(100), label_colors = NULL, label_palette = "Set3",
  cluster_colors = NULL, cluster_palette = "Paired", ...)
}
\description{
Plot a similarity matrix heatmap
}
\keyword{internal}
