% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMLR.R
\name{SIMLR}
\alias{SIMLR}
\title{SIMLR}
\usage{
SIMLR(X, c, no.dim = NA, k = 10, if.impute = FALSE, normalize = FALSE,
  cores.ratio = 1)
}
\arguments{
\item{X}{todo}

\item{c}{todo}

\item{no.dim}{todo}

\item{k}{todo}

\item{if.impute}{todo}

\item{normalize}{todo}

\item{cores.ratio}{todo}
}
\value{
todo
}
\description{
perform the SIMLR clustering algorithm
}
\examples{
SIMLR(X=test$in_X, c=test$n_clust)

}

