% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMLR_Feature_Ranking.R
\name{SIMLR_Feature_Ranking}
\alias{SIMLR_Feature_Ranking}
\title{SIMLR Feature Ranking}
\usage{
SIMLR_Feature_Ranking(A, X)
}
\arguments{
\item{A}{an (n x n) similarity matrix by SIMLR}

\item{X}{an (m x n) data matrix of gene expression measurements of individual cells}
}
\value{
a list of 2 elements: pvalues and ranking ordering over the n covariates 
as estimated by the method
}
\description{
perform the SIMLR feature ranking algorithm. This takes as input the original 
input data and the corresponding similarity matrix computed by SIMLR
}
\examples{
SIMLR_Feature_Ranking(A = BuettnerFlorian$results$S, X = BuettnerFlorian$in_X)

}

