% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.simlr.R
\name{sort.rows}
\alias{sort.rows}
\title{Sort each row of X and return the sorted matrix and the ordering vectors}
\usage{
\method{sort}{rows}(X, cl)
}
\description{
Sort each row of X and return the sorted matrix and the ordering vectors
}
\keyword{internal}
