% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMLR.R
\name{summarise_SIMLR}
\alias{summarise_SIMLR}
\title{Summarise the results of the SIMLR algorithm}
\usage{
summarise_SIMLR(res, .data, data_set, output_dir = file.path("output",
  data_set), max_intermediaries = 6, max_samples = 300)
}
\description{
Summarise the results of the SIMLR algorithm
}
\keyword{internal}
