% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZeiselAmit}
\alias{ZeiselAmit}
\title{test dataset for SIMLR large scale}
\format{gene expression measurements of individual cells}
\source{
Zeisel, Amit, et al. "Cell types in the mouse cortex and hippocampus revealed by single-cell RNA-seq." Science 347.6226 (2015): 1138-1142.
}
\usage{
data(ZeiselAmit)
}
\value{
list of 6: 
	in_X = input dataset as an (m x n) gene expression measurements of individual cells,
 	n_clust = number of clusters (number of distinct true labels), 
 	true_labs = ground true of cluster assignments for each of the n_clust clusters, 
 	seed = seed used to compute the results for the example,
 	results = result by SIMLR for the inputs defined as described,
 	nmi = normalized mutual information as a measure of the inferred clusters compared to the true labels
}
\description{
example dataset to test SIMLR large scale from the work by Zeisel, Amit, et al.
}

