% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.simlr.R
\name{zero_vec}
\alias{zero_vec}
\title{Return a vector of the same length as x, where all elements that are less
than the k'th largest of x are 0. Other elements are untouched.}
\usage{
zero_vec(x, k)
}
\description{
Return a vector of the same length as x, where all elements that are less
than the k'th largest of x are 0. Other elements are untouched.
}
\keyword{internal}
