% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.CIMLR.R
\name{CIMLR}
\alias{CIMLR}
\title{This function has been removed and it is now provided
by the CIMLR package, please refer to https://github.com/danro9685/CIMLR}
\usage{
CIMLR(X, c, no.dim = NA, k = 10, cores.ratio = 1)
}
\arguments{
\item{X}{a list of multi-omic data each of which is an (m x n) data matrix of measurements of cancer patients}

\item{c}{number of clusters to be estimated over X}

\item{no.dim}{number of dimensions}

\item{k}{tuning parameter}

\item{cores.ratio}{ratio of the number of cores to be used when computing the multi-kernel}
}
\value{
clusters the patients based on CIMLR and their similarities
}
\description{
perform the CIMLR clustering algorithm
}
\examples{
   
\dontrun{
CIMLR(X = GliomasReduced$in_X, c = 3, cores.ratio = 0)   
}    
}
