% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMLR_Estimate_Number_of_Clusters.R
\name{CIMLR_Estimate_Number_of_Clusters}
\alias{CIMLR_Estimate_Number_of_Clusters}
\title{CIMLR Estimate Number of Clusters}
\usage{
CIMLR_Estimate_Number_of_Clusters(all_data, NUMC = 2:5,
  cores.ratio = 1)
}
\arguments{
\item{all_data}{is a list of multi-omic data each of which is an (m x n) data matrix of measurements of cancer patients}

\item{NUMC}{vector of number of clusters to be considered}

\item{cores.ratio}{ratio of the number of cores to be used when computing the multi-kernel}
}
\value{
a list of 2 elements: K1 and K2 with an estimation of the best clusters (the lower 
values the better) as discussed in the original paper of SIMLR
}
\description{
estimate the number of clusters by means of two huristics as discussed in the CIMLR paper
}
\examples{

CIMLR_Estimate_Number_of_Clusters(GliomasReduced$in_X,
   NUMC = 2:5,
   cores.ratio = 0)

}
