% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.diffusion.R
\name{dn}
\alias{dn}
\title{Normalizes a symmetric kernel w}
\usage{
dn(w, type)
}
\arguments{
\item{:}{w The symmetric kernel}

\item{:}{type The normalisation type}
}
\description{
Normalizes a symmetric kernel w
}
