% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.diffusion.R
\name{dn}
\alias{dn}
\title{Normalizes a symmetric kernel w.}
\usage{
dn(W, type)
}
\arguments{
\item{:}{w The symmetric kernel}

\item{:}{type The normalisation type
\enumerate{
  \item 'ave' Scales the rows by the column sums
  \item 'gph' Scales each element by the square root of its column and row sums
}}
}
\description{
Normalizes a symmetric kernel w.
}
\keyword{internal}
