% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.diffusion.R
\name{dominate.set}
\alias{dominate.set}
\title{Compute the dominate set for the matrix aff.matrix and NR.OF.KNN}
\usage{
dominate.set(aff.matrix, NR.OF.KNN)
}
\description{
Compute the dominate set for the matrix aff.matrix and NR.OF.KNN
}
