% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.multiple.kernel.R
\name{multiple.kernel}
\alias{multiple.kernel}
\title{Compute and returns the multiple kernel distances}
\usage{
multiple.kernel(x, cores.ratio = 1, calc.dists = TRUE)
}
\arguments{
\item{x}{The data (samples x features)}

\item{cores.ratio}{Proportional of all possible cores - 1 to use.}

\item{calc.dists}{Normalise kernels and convert to kernel distances}
}
\description{
Compute the kernel distances for the hard-wired range sigma (bandwidth scaling) and k (nearest-neighbours)
}
\keyword{internal}
