% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.multiple.kernel.R
\name{multiple.unnorm.kernels}
\alias{multiple.unnorm.kernels}
\title{Compute and returns multiple unnormalised kernels}
\usage{
multiple.unnorm.kernels(x, cl, dist_power = 2)
}
\arguments{
\item{x}{The data (samples x features)}

\item{cl}{Cluster to run parallel tasks on}

\item{dist_power}{The power to raise the distances}
}
\description{
Compute the kernels for the hard-wired range sigma (bandwidth scaling) and k (nearest-neighbours)
}
\keyword{internal}
