% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.diffusion.R
\name{network.diffusion}
\alias{network.diffusion}
\title{Perform network diffusion using K neareast neighbours over the network A}
\usage{
network.diffusion(A, k, scale_by_DD = TRUE)
}
\arguments{
\item{A}{The distance matrix}

\item{k}{The number of nearest neighbours to consider}

\item{scale_by_DD}{Scale W by DD (included in standard network.diffusion() but not in network.diffusion.numc())}
}
\description{
This is called similarity enhancement by diffusion in the paper
}
\keyword{internal}
