% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMLR.R
\name{SIMLR}
\alias{SIMLR}
\title{SIMLR}
\usage{
SIMLR(X, c, no.dim = NA, k = 10, if.impute = FALSE, normalize = FALSE,
  cores.ratio = 1)
}
\arguments{
\item{X}{input data of gene expression measurements of individual cells}

\item{c}{number of clusters}

\item{no.dim}{number of dimensions}

\item{k}{tuning parameter}

\item{if.impute}{should I traspose the input data?}

\item{normalize}{should I normalize the input data?}

\item{cores.ratio}{ratio of the number of cores to be used when computing the multi-kernel}
}
\value{
clusters the cells based on SIMLR and their similarities
}
\description{
perform the SIMLR clustering algorithm
}
\examples{
SIMLR(X = BuettnerFlorian$in_X, c = BuettnerFlorian$n_clust, cores.ratio = 0)

}

