% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMLR_Feature_Ranking.R
\name{SIMLR_Feature_Ranking}
\alias{SIMLR_Feature_Ranking}
\title{SIMLR Feature Ranking}
\usage{
SIMLR_Feature_Ranking(A, X)
}
\arguments{
\item{A}{similarity matrix by SIMLR}

\item{X}{input data of gene expression measurements of individual cells}
}
\value{
pvalues and ranking ordering
}
\description{
perform the SIMLR feature ranking algorithm
}
\examples{
SIMLR_Feature_Ranking(A = BuettnerFlorian$results$S, X = BuettnerFlorian$in_X)

}

