% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.diffusion.R
\name{transition.fields}
\alias{transition.fields}
\title{Compute the transition field of the given matrix}
\usage{
transition.fields(W)
}
\description{
For a given adjacency matrix W, calculate the diffusion operator M <- inv(D) %*% W.
}
\keyword{internal}
